Reflect container now stores more information about its template arguments if a struct/class has any.
Added some more profile scopes.
After parsing all files, 'using'/'typdefs' are parsed. If a type is defined with 'using'/'typedef' and it can't be found as a object which has been parsed, then all variations of the using type with namespaces is added to the alias map. This is to hopefully catch any use case and relate that to the full path with namespaces.
Example: namespace A { namespace B { struct C { }; using D = C; } }, as the struct C has not been reflect and so doesn't exists within the reflect system, the reflect system can't use the full namespace the find it. Instead the alias table will add all variations of C with its namespaces (C, B::C, A::B::C) and point them to A::B::C in the hopes that this will resolve not knowing the full path with namespaces. If there is a conflict then a message is printed to the log to inform the used.